workout_tracker/workout_tracker/src/workout_tracker/models.py
from __future__ import annotations

from datetime import date, datetime

from flask_login import UserMixin
from werkzeug.security import check_password_hash, generate_password_hash

from .db import db


class User(UserMixin, db.Model):
    __tablename__ = "users"

    id = db.Column(db.Integer, primary_key=True)
    email = db.Column(db.String(255), unique=True, nullable=False, index=True)
    password_hash = db.Column(db.String(255), nullable=False)
    display_name = db.Column(db.String(120))
    created_at = db.Column(db.DateTime, default=datetime.utcnow, nullable=False)

    def set_password(self, password: str) -> None:
        self.password_hash = generate_password_hash(password)

    def check_password(self, password: str) -> bool:
        return check_password_hash(self.password_hash, password)

    @property
    def first_name(self) -> str:
        if self.display_name:
            return self.display_name.split()[0]
        return self.email.split("@")[0]

    def __repr__(self) -> str:
        return f"<User {self.id} {self.email}>"


class Workout(db.Model):
    __tablename__ = "workouts"

    id = db.Column(db.Integer, primary_key=True)
    user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False, index=True)
    date = db.Column(db.Date, nullable=False, default=date.today)
    exercise = db.Column(db.String(120), nullable=False, index=True)
    sets = db.Column(db.Integer, default=0, nullable=False)
    reps = db.Column(db.Integer, default=0, nullable=False)
    weight = db.Column(db.Float, default=0.0, nullable=False)
    notes = db.Column(db.Text)
    created_at = db.Column(db.DateTime, default=datetime.utcnow, nullable=False)
    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow, nullable=False)

    user = db.relationship("User", backref=db.backref("workouts", lazy="dynamic"))

    @property
    def volume(self) -> float:
        return float(self.sets * self.reps * self.weight)
