workout_tracker/workout_tracker/requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask_login==0.6.3
python-dotenv==1.0.1
gunicorn==22.0.0
